requirements.txt
asgiref
blinker
Brotli
certifi
cffi
charset-normalizer
click
cssselect2
Django
django-admin-interface
django-colorfield
django-decorator-include
django-decouple
django-widget-tweaks
et_xmlfile
fonttools
idna
itsdangerous
Jinja2
lxml
MarkupSafe
numpy
openpyxl
pandas
pillow
pycparser
pydyf
pyphen
python-dateutil
python-docx
python-slugify
pytz
requests
sib-api-v3-sdk
six
sqlparse
text-unidecode
tinycss2
tinyhtml5
typing_extensions
tzdata
urllib3
weasyprint
webencodings
Werkzeug
zopfli
